added routes and nested routing
