+++ src/App.css
@@ -46,9 +46,9 @@ font-size: 32px;
   position: absolute;
   top: 10px;
   right: 10px;
-  background-color: #fff;
+  /* background-color: #fff;
   padding: 5px;
-  border-radius: 50%;
+  border-radius: 50%; */
 }
 .AddToCartBtn{
   padding: 10px 20px;
