+++ src/Pages/AllProductsPage.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+
+const AllProductsPage = () => {
+  return (
+    <>
+    <div className="allPagesDiv">
+     <h2>All Products</h2>
+    </div>
+
+    </>
+  )
+}
+
+export default AllProductsPage
