+++ src/Pages/CartPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const CartPage = () => {
+  return (
+    <>CartPage</>
+  )
+}
+
+export default CartPage
