+++ src/Pages/WishlistPage.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const WishlistPage = () => {
+  return (
+    <>WishlistPage</>
+  )
+}
+
+export default WishlistPage
